refactor(store): Use store for navbar Closes #1399
