+++ packages/geoview-core/src/core/components/details/details-style.ts
