+++ packages/geoview-core/src/core/components/details/details.tsx
