+++ packages/geoview-core/src/geo/map/map-viewer.ts
@@ -22,7 +22,7 @@ import { NavbarButtons } from '@/core/components/nav-bar/nav-bar-buttons';
 import { FooterTabsApi } from '@/core/components/footer-tabs/footer-tabs-api';
 import { LegendApi } from '@/core/components/legend/legend-api';
 import { Legend2Api } from '@/core/components/legend-2/legend-api';
-import { DetailsAPI as DetailsAPIFooter } from '@/core/components/details-1/details-api';
+import { DetailsApi } from '@/core/components/details/details-api';
 import { DataTableApi } from '@/core/components/data-table/data-table-api';
 import { GeoviewRenderer } from '@/geo/renderer/geoview-renderer';
 import { Select } from '@/geo/interaction/select';
@@ -85,7 +85,7 @@ export class MapViewer {
   legend2!: Legend2Api;
 
   // used to access the details in footer
-  detailsFooter!: DetailsAPIFooter;
+  details!: DetailsApi;
 
   // used to access the data table api
   dataTable!: DataTableApi;
@@ -148,7 +148,7 @@ export class MapViewer {
     this.footerTabs = new FooterTabsApi(this.mapId);
     this.legend = new LegendApi(this.mapId);
     this.legend2 = new Legend2Api(this.mapId);
-    this.detailsFooter = new DetailsAPIFooter(this.mapId);
+    this.details = new DetailsApi(this.mapId);
     this.dataTable = new DataTableApi(this.mapId);
 
     this.modal = new ModalApi(this.mapId);
